% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_descriptions.R
\docType{data}
\name{BlackRat}
\alias{BlackRat}
\title{Black Rat shape dataset}
\format{
A list of 2 objects:
\describe{
  \item{LMArray}{Specimen landmark data in array form, where each row is a landmark, each column a landmark dimension and each slice is an individual specimen. These landmarks have been scaled.}
}
}
\source{
\url{https://doi.org/10.1007/s10914-017-9423-8}
}
\usage{
BlackRat
}
\description{
A geometric morphometric derived dataset containing an data from 100 specimens
with 120 landmarks all collected from Mainland Southeast Asia.
}
\keyword{datasets}
