% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDAOptimisationPar_KFold_functions.R
\name{LDACVManyStepwisePar}
\alias{LDACVManyStepwisePar}
\title{Stepwise LDA correct cross-validation with equal sample size using parallel processing}
\usage{
LDACVManyStepwisePar(
  DiscriminationData,
  GroupMembership,
  EqualIter = 100,
  KFold = 5,
  SampleSize = NA,
  Verbose = FALSE,
  ShapeGPA = FALSE,
  Sliding = NULL,
  SizeShape = FALSE,
  PClim = 10,
  PlotResults = TRUE,
  TestTraining = FALSE
)
}
\arguments{
\item{DiscriminationData}{A matrix of numeric data or an array for shape data. If shape data is used please set ShapeGPA=TRUE. A shape data array is expected to include landmarks as rows, landmark dimensions (either 2 or 3) as columns and specimens as slices. If a matrix is used, please ensure no columns have erroneously been loaded as factors.}

\item{GroupMembership}{A vector of group classifications, either as factors or as characters.}

\item{EqualIter}{an integer determining the number of times the data will be resampled to equal sample size. Default is set to 100.}

\item{KFold}{value default set to 5}

\item{SampleSize}{the sample size to be used. If set to NA the resampling exercise will automatically resample groups to the sample size of the smallest group.}

\item{Verbose}{logical (either TRUE or FALSE, default set to FALSE) to determine whether to return the identifications of each round of resampling (Verbose=TRUE) or to just return the correct-cross-validation percentages of each resampling exercise (Verbose=FALSE).}

\item{ShapeGPA}{logical (either TRUE or FALSE, default set to FALSE) to indicate whether the data is shape data and if so each subset of individuals from the resampling procedure will be processed through a Generalised Procrustes Analysis (GPA) and subsequent Principal Component Analysis (PCA) using the \code{\link[Morpho]{procSym}} function in the package \code{Morpho}.}

\item{Sliding}{if ShapeGPA is TRUE and the shape data has sliding landmarks then a list should be provided here of landmarks to be slid (see function description of \code{\link[Morpho]{procSym}})}

\item{SizeShape}{logical (either TRUE or FALSE, default set to FALSE). If ShapeGPA is TRUE and you wish to analyses form (i.e. shape+log size) then this argument will be passed to the \code{\link[Morpho]{procSym}} and the balanced LDA will be applied to form PCs (see function description of \code{\link[Morpho]{procSym}}).}

\item{PClim}{integer determining the number of PCs to be used in the case that ShapeGPA or PCA are set to TRUE. Default is arbitrarily set to 10.}

\item{PlotResults}{logical (either TRUE or FALSE, default set to TRUE) to determine whether to plot the results of the stepwise discriminant analyses.}

\item{TestTraining}{default set to FALSE, indicates whether an ANOVA and MANOVA should be carried at each step and iteration.}
}
\value{
Returns a matrix of the leave-one-out classifications for all the specimens along with their known classification.
}
\description{
This function takes a matrix of variables of known group membership and returns
the results of a leave-one-out correct cross validation identification for each
specimen to provide a correct cross-validation percentage. This function is for
for stepwise testing of each set of consecutive PCs with resampling to equal sample
size at each incremental increase.
}
\details{
The function is primarily for use with resampling unequal groups to equal sample
size a set number of times. This process is carried out with parrallel processing.
If shape data is used the function offers the option to carry out a new GPA and
subsequent PCA for each resampling exercise (set ShapeGPA=TRUE). If raw data
is used the funciton offers the option to carry out a new PCA with each resampling
exercise. In both these instances where a fresh PCA is carried out the function
will call on the input value in PClim to determine the number of PCs to use.
}
\author{
Ardern Hulme-Beaman
}
\keyword{Geometric}
\keyword{distances}
\keyword{morphometric}
\keyword{shape}
