% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDAOptimisationPar_functions.R
\name{LDACVStepwisePar}
\alias{LDACVStepwisePar}
\title{Stepwise LDA correct cross-validation with equal sample size using parallel processing}
\usage{
LDACVStepwisePar(
  DiscriminationData,
  GroupMembership,
  EqualIter = 100,
  SampleSize = NA,
  Verbose = FALSE,
  ShapeGPA = FALSE,
  Sliding = NULL,
  SlidingLMindex = NULL,
  Bending = TRUE,
  SizeShape = FALSE,
  PClim = 10,
  PlotResults = TRUE,
  CombinePlots = FALSE
)
}
\arguments{
\item{DiscriminationData}{A matrix of numeric data or an array for shape data. If shape data is used please set ShapeGPA=TRUE. A shape data array is expected to include landmarks as rows, landmark dimensions (either 2 or 3) as columns and specimens as slices. If a matrix is used, please ensure no columns have erroneously been loaded as factors.}

\item{GroupMembership}{A vector of group classifications, either as factors or as characters.}

\item{EqualIter}{an integer determining the number of times the data will be resampled to equal sample size. Default is set to 100.}

\item{SampleSize}{the sample size to be used. If set to NA the resampling exercise will automatically resample groups to the sample size of the smallest group.}

\item{Verbose}{logical (either TRUE or FALSE, default set to FALSE) to determine whether to return the identifications of each round of resampling (Verbose=TRUE) or to just return the correct-cross-validation percentages of each resampling exercise (Verbose=FALSE).}

\item{ShapeGPA}{logical (either TRUE or FALSE, default set to FALSE) to indicate whether the data is shape data and if so each subset of individuals from the resampling procedure will be processed through a Generalised Procrustes Analysis (GPA) and subsequent Principal Component Analysis (PCA) using the \code{\link[Morpho]{procSym}} function in the package \code{Morpho}.}

\item{Sliding}{if ShapeGPA is TRUE and the shape data has sliding landmarks then a list should be provided here with vectors of landmarks to be slid (see argument description for outlines in the \code{\link[Morpho]{procSym}} function of the \code{\link[Morpho]{Morpho}} package).}

\item{SlidingLMindex}{a vector of the landmark numbers that represent sliding landmarks (see argument description for SMvector in the \code{\link[Morpho]{procSym}} function of the \code{\link[Morpho]{Morpho}} package).}

\item{Bending}{default set to TRUE. This is passed to the \code{Morpho} package functions to determine if minimum bending energy (TRUE) or minimum procrustes distance (set to FALSE) is used for sliding the landmarks.}

\item{SizeShape}{logical (either TRUE or FALSE, default set to FALSE). If ShapeGPA is TRUE and you wish to analyses form (i.e. shape+log size) then this argument will be passed to the \code{\link[Morpho]{procSym}} and the balanced LDA will be applied to form PCs (see function description of \code{\link[Morpho]{procSym}}).}

\item{PClim}{integer determining the number of PCs to be used in the case that ShapeGPA or PCA are set to TRUE. Default is arbitrarily set to 10.}

\item{PlotResults}{logical (either TRUE or FALSE, default set to TRUE) to determine whether to plot the results of the stepwise discriminant analyses.}

\item{CombinePlots}{logical set to FALSE to indicate whether the results of this analysis should be plotted on a previous result that's currently open (e.g. for randomised analyses).}
}
\value{
Returns a matrix of the leave-one-out classifications for all the specimens along with their known classification.
}
\description{
This function takes a matrix of variables of known group membership and returns
the results of a leave-one-out correct cross validation identification for each
specimen to provide a correct cross-validation percentage. This function is for
for stepwise testing of each set of consecutive PCs with resampling to equal sample
size at each incremental increase following Baylac and Friess 2005.
}
\details{
The function is primarily for use with resampling unequal groups to equal sample
size a set number of times, following Evin et al. 2013. This process is carried
out with parallel processing. If shape data is used the function offers the
option to carry out a new GPA and subsequent PCA for each resampling exercise
(set ShapeGPA=TRUE). If raw data is used the function carries out a new PCA
with each resampling exercise. In both these instances where a fresh PCA is carried
out the function will call on the input value in PClim to determine the number
of PCs to use.

Note that the method in this function is to remove data leakage from unequal sample
size, but that the leave-one-out method here does not train the identification model
separately for each specimen. Instead it is assumed that the single specimen being
left out in the cross validation procedure contributes so little data leakage as to
be unimportant. However, if this is a concern please use the k-fold method in
\code{LDACVManyPar} and \code{LDACVManyStepwisePar} and set k to the sample size of
the smallest sample size.
}
\section{Citations}{


Evin, A., Cucchi, T., Cardini, A., Vidarsdottir, U.S., Larson, G. and Dobney, K., 2013. The long and winding road: identifying pig domestication through molar size and shape. Journal of Archaeological Science, 40(1), pp.735-743.

Baylac and Friess, 2005 Fourier descriptors, procrustes superimposition and data dimensionality: an example of cranial shape analysis in modern human populations D.E. Slice (Ed.), Modern Morphometrics in Physical Anthropology, Springer-Verlag, New York (2005), pp. 145-165
}

\author{
Ardern Hulme-Beaman
}
