% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictUnknowns2EqualGroupsPar_functions.R
\name{PredictUnknownsEqualPar}
\alias{PredictUnknownsEqualPar}
\title{Predict unknown specimens using LDA applied to PCA with resampling to equal sample size using parallel processing}
\usage{
PredictUnknownsEqualPar(
  TrainingData,
  UnknownData,
  GroupMembership,
  EqualIter = 100,
  SampleSize = NA,
  ShapeGPA = FALSE,
  Sliding = NULL,
  SlidingLMindex = NULL,
  Bending = TRUE,
  SizeShape = FALSE,
  PClim = 10
)
}
\arguments{
\item{TrainingData}{matrix where specimens are rows. If shape data (ShapeGPA=TRUE) then the data is expected in an array format where rows are landmarks, columns are landmark dimensions and slices are specimens. The order of the specimens should match the GroupMembership vector.}

\item{UnknownData}{as for Training data but with the unknown data.}

\item{GroupMembership}{A vector of group classifications, either as factors or as characters.}

\item{EqualIter}{an integer determining the number of times the data will be resampled to equal sample size. Default is set to 100.}

\item{SampleSize}{the sample size to be used. If set to NA the resampling exercise will automatically resample groups to the sample size of the smallest group.}

\item{ShapeGPA}{logical (either TRUE or FALSE, default set to FALSE) to indicate whether the data is shape data and if so each subset of individuals from the resampling procedure will be processed through a Generalised Procrustes Analysis (GPA) and subsequent Principal Component Analysis (PCA) using the \code{\link[Morpho]{procSym}} function in the package \code{Morpho}.}

\item{Sliding}{if ShapeGPA is TRUE and the shape data has sliding landmarks then a list should be provided here with vectors of landmarks to be slid (see function description of \code{\link[Morpho]{procSym}}).}

\item{SlidingLMindex}{a vector of the landmark numbers that represent sliding landmarks.}

\item{Bending}{default set to TRUE. This is passed to the \code{Morpho} package functions to determine if minimum bending energy (TRUE) or minimum procrustes distance (set to FALSE) is used for sliding the landmarks.}

\item{SizeShape}{logical (either TRUE or FALSE, default set to FALSE). If ShapeGPA is TRUE and you wish to analyses form (i.e. shape+log size) then this argument will be passed to the \code{\link[Morpho]{procSym}} and the balanced LDA will be applied to form PCs (see function description of \code{\link[Morpho]{procSym}}).}

\item{PClim}{integer determining the number of PCs to be used in the case that ShapeGPA or PCA are set to TRUE. Default is arbitrarily set to 10.}
}
\value{
Returns a matrix of the leave-one-out classifications for all the specimens along with their known classification.
}
\description{
This function is designed for predicting unknown specimens to group by
carrying out an LDA on reduced dimensions by PCA. The function resamples
reference groups (aka training groups) to equal sample size and carries out
PCA on the reference groups. The unknown specimens (aka test specimens) are
then projected into the same PCA space. This ensures there is no data leakage
into the identification model from either the unequal reference group sample
sizes or from the unknown specimens. In other words, if the reference groups
are unequal then the larger groups will dispropotionately influence the PCA
upon which later predictive models will act on, this is known as data leakage
and is avoided using this function.
}
\details{
The function can also be applied to shape data in which case it expects an
array of data with landmarks as rows, dimensions as columns and specimens as
sliced (for shape data set ShapeGPA=TRUE). In the case of shape data, both a GPA
and a PCA are carried out for each resampling excercise to equal sample sizes,
unknown specimens are then projected into the shape space defined by that equal
sample size and then those rotated specimens a projected into the PCA space.

This function is for application to a set number of PCs once the optimal number
is established via correct cross-validation. To establish the optimal number
of PCs required see the function \code{LDACVManyStepwisePar} for stepwise
testing of each set of consecutive PCs.
}
\author{
Ardern Hulme-Beaman
}
